Загрузка данных из бд.

+++ src/history/candle.rs
@@ -7,11 +7,11 @@ pub struct Candle {
     high: f64,
     low: f64,
     close: f64,
-    volume: u64,
+    volume: u32,
 }
 
 impl Candle {
-    pub fn new(date_time: DateTime<Local>, o: f64, h: f64, l: f64, c: f64, v: u64) -> Candle {
+    pub fn new(date_time: DateTime<Local>, o: f64, h: f64, l: f64, c: f64, v: u32) -> Candle {
         return Candle {
             dt: date_time,
             open: o,
